+++ app/controllers/folio/console/dashboard_controller.rb
@@ -3,7 +3,13 @@
 module Folio
   class Console::DashboardController < Console::BaseController
     def index
-      redirect_to console_nodes_path
+      if path = Rails.application.config.folio_console_dashboard_redirect
+        if respond_to?(path)
+          redirect_to send(path)
+        elsif main_app.respond_to?(path)
+          redirect_to main_app.send(path)
+        end
+      end
     end
   end
 end